Cant really remember what I did here
But in the end it was to draw the DefFunc picture on the heap. I think
it might be better to just show the iconified version of the objects
on the heap, because it looks weird to have for example a function
body array on the heap drawn, and then drawn again when the actual
function definition is drawn.