Selectable items in the heap
Implemented a pretty bad way of selecting elements in the heap. I think
a better way would be for the heap graphical element to subscribe to
events (up and down keys), and handle everything in the heap graphical
class. Advantage there is other graphical elements can decide what they
want to subscribe to.

Flipside though is keeping the interface coherent. In a way having the
Boxer::Screen object deciding what keys do what is also good. Something
has to have an overview of how the interface behaves. By moving the
signal registration into the heap graphical class, I am saying that each
element then is deciding how it behaves in response to input. That could
lead to elements behaving in inconsistent ways.

I could introduce a new class, whos job it is to define how the entire
application interface is layed out: where on the screen the heap sits
(or is the heap even visible), what keys do what and when etc.